yes
What is the FitnessGram PACER Test?
FitnessGram
PACER Test

The FitnessGram PACER Test is a multistage aerobic capacity test that progressively gets more difficult as it continues.

The test is used to measure a student's aerobic capacity as part of the FitnessGram assessment. Students run back and forth as many times as they can, each lap signaled by a beep sound. The test get progressively faster as it continues until the student reaches their max lap score.

The PACER Test score is combined in the FitnessGram software with scores for muscular strength, endurance, flexibility and body composition to determine whether a student is in the Healthy Fitness Zone™ or the Needs Improvement Zone™.
New Sound. New Beeps. Same Test.

For nearly three decades, the Progressive Aerobic Cardiovascular Endurance Run (PACER) Test has been used as part of the FitnessGram assessment in thousands of P.E. classes in all 50 states.  Now the legendary voice of the school fitness test is changing. The Cooper Institute is proud to announce the launch of the FitnessGram PACER Test Remixes! powered by Hip Hop Public Health.

The new album is a collaboration between The Cooper Institute and Hip Hop Public Health, two organizations dedicated to improving youth fitness, reducing childhood obesity, and promoting life-long healthy habits through research and education.

FitnessGram and Hip Hop Public Health

The six new tracks revitalize the decades-old fitness test with a fusion of hip-hop, pop, electronic dance music, and Latin-inspired beats, bringing a DJ dance party vibe to the gym.

The iconic voice of the original test has been replaced by both male and female voices in English and Spanish to motivate and encourage participants throughout each of the 22-minute tracks.

The combination of diverse voices and culturally relevant beats is designed to inspire a new generation of students to get active and stay healthy now and well into adulthood.
The classic sound of the original 20-Meter FitnessGram PACER Test is now available with visual cues on YouTube!